Solve Exercise 2-02: No logical operators
The text doesn't specify which technique to use. The section outlined
operator precedence, but I couldn't think of a way to check for truth in
the three main expressions without logical operators. So I turned to a
recursive function. It works, but I'm not sure if it's what K&R were
after.