Re-solve Exercise 1-16: arbitrarily long maxline
A completely rewritten solution which follows the spec given in the book
exactly; make the program work for arbitrarily long input lines while
modifying `main()` only. I've only changed the name of the macro MAXLINE
to BUFFSIZE for clarity - otherwise the code is typed up exactly as
found in the book.

It demonstrates the use of a buffer without modifying any of the helper
functions.