1-16 solution code and comment style cleanup
Also of note is the length that the program reports includes the newline
character at the end. This has been clarified in a comment. The exercise
text does not specify whether to include newlines as part of the length,
so it was left alone.

Removed the contributor notice, since contributor acknowledgements are
centralized.