Fix Exercise 5-01 to put invalid signchar back on buffer
A nitpick at the specs: getint() is supposed to put an invalid sign character
(i.e. a '+' or '-' not followed by a digit) back on the buffer, presumably
so the input can be used elsewhere.

I also print the contents of the buffer when the program exits so its
correctness can be verified.